Fix recursive CTE mergejoin having motion on WTS
For a recursive CTE, planner currently creates a mergejoin plan which
moves the WorkTableScan results to a singleQE using
CdbPathLocus_MakeSingleQE
That would introduce motion in between Union and WorkTableScan. The
WorkTableScan results should be local to each segment and shouldn't be
broadcasted/moved to other segments. We shouldn't be allowing any motion
in between Union and WorkTableScan. This commit adds a check for the
last resort case to avoid adding that motion. If there is a WTS on any
side in the last resort case, we should just fail (no possible plan
using mergejoin)

Added simplified test to force a mergejoin for such recursive CTE, the
test shouldn't create a mergejoin plan.

Co-authored-by: Zhenghua Lyu <zlv@pivotal.io>
Co-authored-by: Alexandra Wang<lewang@pivotal.io>