Disallow non-standalone ALTER distribution
If AT SET DISTRIBUTED BY, or AT SET WITH (reorganize=true|false) is run
today along with another subcommand, it will result in ERRORs such as:

ALTER TABLE foo ADD COLUMN c int, SET DISTRIBUTED BY(b);
ERROR: attribute number 3 exceeds number of columns 2

This kind of error results when the CTAS for the temp table is
dispatched as part of ATExecSetDistributedBy(). Since SET DISTRIBUTED BY
is run in the last pass, the ADD COLUMN takes effect on the QD and is
considered as part of the dispatched QueryDesc for the CTAS. However,
since the AT command has not been dispatched yet to the QEs (which only
happens after ATController()), the QE hasn't yet applied the results of
the ADD COLUMN operation.  Thus, the QEs cannot reconcile the reference
to the added column, and an ERROR results.

Other ERRORs can also result from having an inconsistent view of the
catalog on the QEs in the CTAS.

Supporting AT SET DISTRIBUTED BY with other subcommands would require
abandoning the current approach, and aligining it with regular AT
workflow. Since, that is a larger change, we ban such usage at the
moment.

Co-authored-by: Andrew Repp <reppa@vmware.com>