[7X]Fix utilities do not honor -d flag when COORDINATOR_DATA_DIRECTORY is not set. (#16433)
* Fix: utilities do not honor -d flag when COORDINATOR_DATA_DIRECTORY is not set.

Issue: Following are the gpMgmt utilities that do not honor the -d flag, when
COORDINATOR_DATA_DIRECTORY is not set.
1. gpstart
2. gpstop
3. gpstatus
4. gprecoverseg
5. gpaddmirror

RCA: to get the coordinator data directory gp.getcoordinator_datadir()
function is called from the above-listed utilities. the function does not
have any provision to return the coordinator data directory which is
provided with the -d flag. currently, it looks for COORDINATOR_DATA_DIRECTORY
and MASTER_DATA_DIRECTORY env variable.

    also in some of the utilities we were creating lock files before
parsing the provided options which looks like the design flow that was
causing the utilities to crash when looking for coordinator data
directory.

Fix: Added a global flag which holds the data directory provided with -d
option. so when we run the utility and do parsing it sets the flag with
the provided datadirectory and the same will be returned when we call
gp.gp.getcoordinator_datadir().

Test:
Added behave test cases to use the provided data directory if COORDINATOR_DATA_DIRECTORY
is not set.
Added behave test case to check if the provided coordinator data directory
is preferred over the already set coordinator_data_dir env variable.
it is tested by setting a wrong COORDINATOR_DATA_DIRECTORY env variable
and when we run the utility with the correct data directory using the -d option then
the utility should execute successfully.

(cherry picked from commit 5adbaa65073511925e191d8d3da83d212f4f4e5f)