Fix flaky test gp_tablespace_with_faults
The test was flaky with a diff like:

```
@@ -397,9 +397,9 @@
 (8 rows)

 select list_tablespace_dbid_dirs(:expected_number_of_tablespaces, :'tablespace_location');
-                                 list_tablespace_dbid_dirs
+                                                 list_tablespace_dbid_dirs
 ------------
- {Success,"expected 1 tablespaces","found 1 tablespaces",/tmp/my_tablespace_for_testing/2}
+ {Failed,"expected 1 tablespaces","found 2 tablespaces",/tmp/my_tablespace_for_testing/2,/tmp/my_tablespace_for_testing/5}
 (1 row)
```

The reason is that there's a chance the mirror has already replayed the WAL to
create tablespace, so we can have two orphaned tablespace directories instead of
one. Now make it panic even before we write the WAL to create tablespace to make
it not flaky.

(cherry picked from commit ffd5c6f7a72df09e06e6cbf8767cbb7a419d75de)