Derive additional distribution spec from union all
**Motivation:**

Currently, the distribution spec delivered by union all is linked to its outer
child. This is because we request 'Any' from the outer child and a matching spec
from the other children. To comprehend why this may not always be suitable,
let's examine a few scenarios where this approach falls short.

Eg.1 Union all of (1) small replicated table and (2) large randomly/hash
distributed table

The outer child delivers the `Replicated` spec. To match the outer child's spec,
we conform the inner child into `Replicated` through a broadcast motion.
Broadcasting a large table is very resource intensive. A more performant plan
would be enforcing the smaller outer child to become `Random` through a hash
motion.

Eg.2 Union all of (1) universal input and (2) large randomly/hash distributed
table
The outer child delivers the `Universal` spec. Since `Universal` cannot be
enforced, we convert the inner child to a `Singleton` through a gather motion.
All the data is then executed on the coordinator, causing it to become a
bottleneck and negating the advantages of MPP.

**Implementation:**

This PR resolves the previously mentioned problems by introducing a new
distribution request for the union all operation. It requests the outer child to
be `Non-Singleton`, and the other children to match the spec delivered by the
outer child. Let's review how this approach tackles the issues above.

In Eg.1, the outer replicated child doesn't satisfy the `Non-Singleton` request
(disabled by `fAllowReplicated`). We enforce it into a `Random` spec through a
duplicate-sensitive random motion. We match the inner child's spec to that
delivered by the outer child. No motion is enforced on the inner child. This
plan prevents the costly broadcasting of a large table.

In Eg.2, the outer universal child doesn't satisfy the `Non-Singleton` request.
Similar to Eg.1, it undergoes a one-time filter based on segment id, and now
delivers the `Random` property. No motion is imposed on the inner child. The
plan leverages the parallelism of scanning a large table across multiple
segments, and prevents the coordinator from becoming the execution bottleneck.

**Discussion:**

It's important to clarify why we don't introduce more distribution requests.

First of all, each optimization request entails an associated costs. We would
only expand the search space if it significantly improves the likelyhood of
finding a better plan. This is why we introduce the <Non-Singleton, match>
request to address suboptimal plans that request multiple union all children to
become 'Singleton' or 'Replicated.'

By this logic, we should avoid requesting 'Singleton' or 'Non-Replicated'. Both
distribution specs enforce through a gather motion, which is generally deferred
to maximize parallelism. We should also refrain from requesting 'Replicated',
which could result in broadcasting large tables. These requests would unlikely
worsen the final plan, but equally unlikely enhance it.

This PR preserves the practice of aligning the distribution requests for the
inner children with the spec delivered by the outer child. This is because the
ORCA cascade framework doesn't allow requests to be tailored based on inherent
properties. Therefore, this PR doesn't offer additional benefit for the
following scenario:

Eg.3 Union all of (1) small randomly/hash distributed table and (2) small
universal input

In this scenario, it would be more cost-effective to align the distribution spec
of the outer child with that of the inner child. This would enable the small
outer table to be gathered onto the coordinator. However, the benefit of this
optimization is limited, as the execution of small tables is usually very brief.

(cherry picked from commit 4bbbe4e49e6ad24ddf8b177cc756cd0c86c27072)