Ignore database binswap_connect in binary swap test
The database binswap_connect that is used for binary swap test is going to be
re-created after we swapped binary. So it will be expected that there could be
difference in the dumping script for that database. Ignore it when dumping.

(cherry picked from commit 68188447e375ef06985aaba806fae8b55a437013)