fix GPDB_96_MERGE_FIXME in create_ctescan_path (#16649)
fixme in create_ctescan_path is as bellow:
```
// GPDB_96_MERGE_FIXME: Why do we set pathkeys in GPDB, but not in Postgres?
// pathnode->pathkeys = NIL;    /* XXX for now, result is always unordered */
pathnode->pathkeys = pathkeys;
```
By now we can not find the commit info of why we use pathkeys in GPDB, but not in Postgres.
The codes "pathnode->pathkeys = pathkeys;" are already introduced after commit: 6b0e52bead. 
And I can not find much more infos in historical repos too.

Sometimes we can reduce extra sort path when the pathkeys of cte node and the pathkeys of subpath 
node are the same.
So we should keep these codes. Just remove GPDB_96_MERGE_FIXME.