Fix ORCA producing incorrect plan when handling SEMI join with RANDOM distributed table (greenplum-db#16611) (#16615)
When ORCA transforms a SEMI join to an INNER join, a LogicalGbAgg is introduced in the inner side of the
INNER join to remove duplicate inner keys. Subsequently, this join is attempted to be transformed into
a LogicalIndexApply. ORCA determines if this transformation can be done by checking if the inner side's
grouping columns contain all distribution columns.

However, when the inner side's distribution columns come from a RANDOM table, the checking always
evaluates to true and allows the transform to proceed. But RANDOM distribution cannot satisfy this
transformation because the same 'distribution' key may appear in different segments.

This commit resolves the issue by adding a check for the size of RANDOM distribution columns and forbidding
the illegal transform.

Co-authored-by: wuyuhao28 <wuyuhao28@github.com>
(cherry picked from commit 1606f347a2e1454d5178198b7ac3379f8998afb6)