Add more functions as NDV preserving to improve join estimates in Orca (#16660)
https://github.com/greenplum-db/gporca/pull/585 introduced funcionality
to improve join estimates when a join condition contains certain
operands that maintain join cardinality. Please see that PR for
relevant background.

Building on that, there's additional operators that preserve NDVs that
would lead to better plans. Note that this is not exhaustive, but are
likely the most common cases.

For example, consider the following join:

test=# explain select * from foo, bar where foo.a=bar.a;                                                                                                                                                                                                                                         QUERY PLAN
---------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..862.18 rows=1000 width=8)
   ->  Hash Join  (cost=0.00..862.15 rows=334 width=8)
         Hash Cond: (foo.a = bar.a)
         ->  Seq Scan on foo  (cost=0.00..431.01 rows=334 width=4)
         ->  Hash  (cost=431.01..431.01 rows=334 width=4)
               ->  Seq Scan on bar  (cost=0.00..431.01 rows=334 width=4)
 Optimizer: GPORCA
(7 rows)

If instead the query is:

test=# explain select * from foo, bar where foo.a+1=bar.a+3;
                                         QUERY PLAN
---------------------------------------------------------------------------------------------
 Hash Join  (cost=0.00..873.61 rows=400000 width=8)
   Hash Cond: ((foo.a + 1) = (bar.a + 3))
   ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..431.02 rows=1000 width=4)
         ->  Seq Scan on foo  (cost=0.00..431.01 rows=334 width=4)
   ->  Hash  (cost=431.02..431.02 rows=1000 width=4)
         ->  Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..431.02 rows=1000 width=4)
               ->  Seq Scan on bar  (cost=0.00..431.01 rows=334 width=4)
 Optimizer: GPORCA
(8 rows)

We should estimate the same number of rows (1000) for this join rather
than the 40% of the cross product we currently do. This is because when
calculating join cardinality, we use the NDVs. For certain operators,
such as adding a constant to an int or timestamp, the NDVs will remain
the same.

(cherry picked from commit 82106c2ad0ce830ee20ad1ae2eff5bff21fe6bc7)