Improve cardinality estimation for projection of ndv-preserving columns (#16664)
When calculating the join cardinality, we lookup the histogram that
corresponds to the colrefs in the join condition. However, when
projecting a column such as int+5, we create a new colref with the new
histogram as a default histogram with no buckets. This means that we
lose all of the information about the underlying colref. When
calculating the join cardinality, we then check if any of the histograms
are empty (no buckets). If so, we set the cardinality as the max rows of
the 2 histograms.

This PR utilizes the work done in
6804d0504d2d8f6d354231e0b73825513774a7bb, and improves the
estimates for ndv-preserving scalar ops such as addition and
subtraction. Now, when projecting a column of a scalar op with a
constant, we'll use the histogram of the underlying colref. The main
useful aspect of these histograms is their NDVs. While the values and
bounds will be different, the NDVs will be the same, and the NDVs are
used in join calculations.

For example, in the query below:
```
explain  select * from date_dim where d_month_seq between (select distinct
d_month_seq+1
from   date_dim where d_year = 1999 and d_moy = 3)
and  (select distinct d_month_seq+12
from   date_dim where d_year = 1999 and d_moy = 3);
```
The projection of `d_month_seq+12` and `d_month_seq+1` are used in the
join filter, `(date_dim.d_month_seq <= ((date_dim_2.d_month_seq +
12)))`. Instead of comparing with empty histograms, we'll have a
histogram using the `d_month_seq` colref.

(cherry picked from commit b0e3251fdbd29bb78f23c0b0a4f2fedc665f5e98)