autostats: Use SKIP_LOCKED for ANALYZE
Ideally, we should not have autostats wait on lock acquisition to help:

* prevent deadlocks - which have been reported in the past
* make the autostats session finish in a more timely and predictable
fashion

Autostats issued ANALYZE will NOT block on lock acquisition. Just like
autoanalyze in 7X, we now use the SKIP_LOCKED option, if
gp_autostats_lock_wait = off (default). If it is on, then waits for the
lock will be performed like before.

Reviewed-by: Huansong Fu <fuhuansong@gmail.com>