Enumerate bushy trees when considering partition selectors
We currently don't enumerate bushy trees in most cases as that can
explode the search space and add to the optimization time. Typically,
bushy trees aren't that great, especially considering motions. However,
one of the common patterns for partition selectors is as follows:

```
HJ
 -HJ
    -PT2
    -PS2
      -Scan with predicate
 -HJ
    -PT1
    -PS1
      -Scan with predicate
```

This bushy tree, where there is a partition selector on the outer side and a selective partition selector on the inner, is common enough that we should explore it.

(cherry picked from commit 4f909484e283d31d3330914f643ef37f1600a387)