Do not throw error for INSERT into a table w/ foreign key constraint
Foreign key constraint is not supported in GPDB. However, the expected behavior
is that we should ignore the constraint and there shouldn't be errors.

But before this commit we would encounter an ERROR when performing INSERT on an
AOCO table that has foreign key constraint:

CREATE TABLE a(id int primary key);
CREATE TABLE b(id int references a) WITH (appendonly=true, orientation=column);
WARNING:  referential integrity (FOREIGN KEY) constraints are not supported in Greenplum Database, will not be enforced

INSERT INTO b VALUES (1);
ERROR:  Trigger is not supported on AOCS yet (trigger.c:2714)  (seg4 #######:16004 pid=3948271) (trigger.c:2714)

The error message, which was for the purpose of disabling row trigger for AO/CO
tables, got in the way of ignoring foreign key constraint. Further investigation
in the code history reveals that the error message was an outdated one which is
no longer needed due to other errors that disable row trigger for AO/CO tables.
Delete the error message now and add test cases.

Fix #16770.