CreateStmt: Introduce notion of origin
This struct member describes the provenance of the CreateStmt, namely
whether it was generated or not as part of GP style classic partitioning
syntax.

Also, it fixes a slight inconsistency where in makePartitionCreateStmt,
we would always indicate that we are part of a GP style alter.

(cherry picked from commit a83c0f841bf1e7f0008c1952fc7f1b42357a6679)