Subquery Scan code cleanup (#16891)
In 7X ORCA, subquery scan is not used. Only the classes for DXL ,parsing of DXL
and translation of DXL to planned statement is present in the code. This came
in with the initial Orca code (back in 2013) and hasn't been touched since.
None of the logic beyond translation was ever implemented. This PR cleans up
the code for subquery scan.

(cherry picked from commit 10abbc1ae2a980666ce187ba866e225a9c5d3201)