[ORCA] Fix missing pdshashedEquiv in IndexOnlyScan (#16898)
Under InnerIndexNLJoin it's also a problem of IndexOnlyScan on AO/AOCS tables
that can not generate IndexScan but can generate IndexOnlyScan.

For example:

  create table test (a int, b int) with (appendonly='true') distributed by (a);
  create index test_a_idx on test(a);
  create index test_b_idx on test(b) include (a);
  insert into test select i,i from generate_series(1, 10000)i;
  analyze test;
  explain (costs off) select t1.a from test t1, test t2
  where t1.a = t2.a and t1.b < 10;
                           QUERY PLAN
  -------------------------------------------------------------
   Gather Motion 3:1  (slice1; segments: 3)
     ->  Nested Loop
           Join Filter: true
           ->  Broadcast Motion 3:3  (slice2; segments: 3)
                 ->  Index Only Scan using test_b_idx on test
                       Index Cond: (b < 10)
           ->  Index Only Scan using test_a_idx on test test_1
                 Index Cond: (a = test.a)
   Optimizer: GPORCA
  (9 rows)

This Broadcast Motion is obviously not necessary, but it appeared because
IndexOnlyScan did not generate pdshashedEquiv just like other index or bitmap
index scan. It's an omission from commit 3b72df1810.

TODO: It has also been noticed that the cost of index only scan is similar to
the cost of index scan, which might be improved upon.

(cherry picked from commit ede206b4639003846e0b673db4c870b8df45c580)