Request syncrep for the forget commit in the remote_apply mode
Currently we lazily flush the forget commit record of a dtx and sync it to a
mirror. This has been fine since dtx recovery can take care of losing the
record and regenerate it. Also, allowing syncrep could have negative performance
impact so it shouldbe limited.

However, in the case of enabling hot standby with synchronous_mode=remote_apply,
it is absolutely required that the forget commit is applied on the standby as
soon as the query is completed on primary. Because otherwise, the standby might
not see the forget commit which makes it think the dtx still in-progress. So it
would view the changes made by the dtx invisible. This would be an incorrect
behavior of remote_apply.

So now request it before supporting hot standby in the subsequent commits.

(cherry picked from commit 5f53e448b92ce182b345820bacc4ce27c4c4933f)