Fix "unrecognized join type" error with LASJ Not-In and network types
During the GP7 merge, separate join selectivity functions were added for network
types. However, GPDB has a separate join, JOIN_LASJ_NOTIN, that does not
exist in Postgres and thus was not added as a possible case.

I followed the same pattern in selfuncs.c, treating it as a semi/anti
join for calculating the selectivity.