Update CDynamicPtrArray::Sort/IsSorted to require CompareFn
Prior to this commit, the default behavior of Sort/IsSorted used a
comparison function that dereferenced the first sizeof(ULONG). The
problem is that for many cases the default compare function doesn't make
sense; and worse is that it hides that from the user. For example,

  ```c++
  CWStringConst *b_string = GPOS_NEW(mp) CWStringConst(mp, "b_string");
  CWStringConst *a_string = GPOS_NEW(mp) CWStringConst(mp, "a_string");

  StringPtrArray *array = GPOS_NEW(mp) StringPtrArray(mp);
  array->Append(b_string);
  array->Append(a_string);

  array->Sort();

  // FAILS!
  GPOS_ASSERT((*array)[0]->Equals(a_string));
  // FAILS!
  GPOS_ASSERT((*array)[1]->Equals(b_string));
  ```

In this case, the first sizeof(ULONG) bytes of CWStringConst does not
correctly influence the sort order. In fact, the string data doesn't
even exist on the object itself.

This commit forces the user to specify a CompareFn. It is still possible
to shoot oneself in the foot, but now it's not the default behavior.

(cherry picked from commit 576811061a97deef9f537f3669d2dc3e74f48d0d)