Fix pg_hint_plan build error with --disable-orca
Commit e781fe0db4 added a hook and commit 9e27c7c36f set the hook. But
the hook was preprocessed out if configured with --disable-orca.

This fix limits the scope of the hook to when orca is configured. It is
admittedly a bit ugly, but necessary in order to preserve ABI
compatibility which prohibits further changing the exported
plan_hint_hook.

(cherry picked from commit 19a4b7e108b9dd9bcc8f62d1a7ff233554155a04)