Support Query Parameters in Orca
This commit adds support for Query Parameters and parameterized queries
to Orca. These come in 2 forms:

1. As functions, eg:
```
CREATE FUNCTION somefunc(INT, INT) RETURNS INT
AS $$ SELECT $1+$2 $$ LANGUAGE SQL STABLE; SELECT * FROM somefunc(3,8);
```

2.And as parameterized queries:
```
PREPARE qry as SELECT * from foo
where a=$1; EXECUTE qry(2);
```

To accomplish this, Orca now parses and passes through a scalar param as
a scalar operator, and then translate this as a Param that GPDB can
understand. This Param is a scalar op just like a CScalarConst, a
ScalarIdent, or any other scalar operator, and fits very nicely into the
Orca framework. The logic works whether the param is in a qual, targetlist,
limit, subquery, etc., and is why there aren't many major logic changes
here--it's mostly just translation.

Some notes:
1. Static partition elimination currently is NOT done with
parameters in Orca. This is because Orca does static elimination at
planning time. Planner handles this by
adding an Append node to the top of the scan, which functions similarly to Orca's
partition selector. In the future, Orca can support this by adding an
Append with a Partition Selector (similar to how 6X does static
elimination) if the Dynamic Scan's qual includes a param.

2. When using parameterized queries (eg: PREPARE..EXECUTE), Orca will
almost always choose a custom plan by default. If plan_cache_mode is set
to force_generic_plan, then it will choose the generic (parameterized)
plan. This is because the cardinality estimates in Orca are using a more
conservative estimate (40%) of the base table. Additionally, Orca's cost
model is not accounting for the planning cost--so the cost of the
generic plan will typically be more than the custom plan and the
optimizer will choose to use the custom plan.

3. Cardinality estimates with query parameters must make assumptions
since we don't know what we're actually comparing against. For example
in `select * from foo where a=$1`, the optimizer must generate an
estimate without knowing the value of $1. In Orca, we currently use the
default unsupported operator value of 40% of the underlying cardinality.
We can make better estimates, for example this should simply be 1/NDV.

This commit also adds and updates some ICW tests. In one case, it
modifies a test to use an explicit sequence instead of serial column.
Using a sequence allows us to specify the cache size which prevents
inconsistent results.

(cherry picked from commit da091d14fab0d0924a5f1e2138e23cc19158a27f)