Fix race between termination and resqueue wakeup
We fix a leak of a resouce queue slot that can be caused by the
following sequence of events:

1. A backend waiting on a resource queue lock receives the TERM signal
and is just about to perform cleanup in:
AbortOutOfAnyTransaction() -> .. -> ResLockWaitCancel()

2. Another backend releases its position in the queue, finds the 1st
backend in the wait queue, evicts it and grants it a slot in the
resource queue.

3. ResLockWaitCancel() resumes for the 1st backend, doesn't find itself
on the wait queue and is a no-op. Thereafter, the backend finishes
termination without cleaning up it's increment set, causing the leak.

Note:
Termination is special in the sense that we don't call ResLockRelease()
in the PG_CATCH() in ResLockPortal()/ResLockUtilityPortal(), as we do
for other exceptional happenstances such as cancellation/deadlock. This is
because we can't catch FATAL errors (like the one that arises from
termination). We never get to clean up the external grant in this case.

To solve this issue, we clean up such dangling increment sets/locks during
PortalDrop() and in AtExitCleanup_ResPortals().

PS: The fact that we have to recover in this fashion is due to our
TRY/CATCH cleanup workflow, something that upstream code advises against
in WaitOnLock(). Fixing that is non-trivial and is another matter.

Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>
Reviewed-by: Xuejing Zhao <zxuejing@vmware.com>
Reviewed-by: Jimmy Yih <jyih@vmware.com>
Reviewed-by: Huansong Fu <fuhuansong@gmail.com>
(cherry picked from commit 723dfb54d8490f7a59d02d9d8efa684c4da5ba8a)