Prevent overwriting of workers in ORCA (#17097)
In some cases, during optimization of queries with call to a function, it is observed that another instance of ORCA optimization is launched.

At the start of optimization, a new worker is registered in the 'Worker Pool Manager'. Currently ORCA supports only one worker, if a new worker is registered on top of an existing worker, then the details of the old worker is lost, causing undefined behaviour. Thus, through this PR, aim is to stop overwriting of workers.

---------

Co-authored-by: nishant sharma <nishants4@nishants46M6RY.vmware.com>

(cherry picked from commit a99e12037c6fc3a9ad9c5b85be98810a63cb6e26)