Fix gpsd test flakes
By explicitly updating pg_statistic, there is the possibility that
autoanalyze analyzes the table between when it is set and when we check
for the expected result. In this case, the intent of the test is to
ensure gpsd dumps quotes. I've added a row with single quotes and
removed the explicit update of pg_statistic.

(cherry picked from commit c0230d50b48174a2c06e7723277832ec3f09a2de)