Require plan hint field in optimizer config
Adding optional fields to the optimizer config is cumbersome and should
not be the norm going forward. Cost model and hint have default values
which we chose to allow to be omitted from the MDP. But, it really
complicates the parsing logic. Let's not make plan hints optional, too.
In the future when we add new fields, we can add it from the end growing
down.

(cherry picked from commit bf184e2f304d07dcf5bb138a9c9549b8681b0bf4)