Consider skew of null values when costing redistributes in Orca (#17311)
When a column contains many null values, distributing on this column
will result in all the tuples going to a single segment, resulting in
computation skew. This commit takes this skew into account when costing
redistribute motions by using the following formula:

skewed_percent: null_fraction + (1/num_segments)
expected_percent_if_uniform: 1/num_segments
skew_ratio = skewed_percent / expected_percent_if_uniform

The greater the null frequency or number of segments, the
greater the skew. We cap this at 10 to prevent significant plan changes.

Consider if a cluster has 2 segments and 40% of the tuples of a column
are null. If all of these tuples go to one segment, it won't be much of
an issue as each segment expects 50% of the tuples. However, if there
are 100 segments, each segment expects 1% of the tuples. With 40% null
values, one segment will get 40% of the tuples and do 40X as much work
as other segments.

There are a good number of plan changes with this. Although the cost
changes aren't very large, I do see redistributes being converted to
broadcasts, and redistributes on different columns. As far as I can
tell, these are good changes--the redistributes previously were on
columns with many null values.

Note, Orca already considers skew if the number of NDVs are less than
the number of segments, which is a narrow use case. Additionally, if the
optimizer_skew_factor guc is set to 1 or greater (0 by default), Orca
will sample and penalize skew. We may want to consider setting this to 1
by default, but it's a more risky approach.

(cherry picked from commit eeaada0f901c744863995e383c84a183241eccce)