Use full precision for statistics values in Orca for mdps (#17386)
This change doesn't have any effect on production code, but brings in
line the behavior between mdps and queries against the database.
Previously, if a frequency or ndv value was very small, it could get
truncated to 0. Although we don't do this translation when querying the
database, it meant that our mdps weren't completely equivalent and could
produce different behavior.

d755bb5 initially introduced greater precision for buckets to prevent
comparison issues, and this commit extends this behavior to frequencies
and ndvs.

Note: CFilterCardinalityTest was modified to not apply the filter twice, since the 
greater precision causes issues with Double. The test takes an input file, applies a filter, and expects an output file. Then it applies the same filter again, and verifies that the output still matches (which intuitively makes sense, data that has already been filtered, we should get the same data no matter how much we reapply the same filter).

Before we truncated at 6 digits, so the loss of precision didn't matter. Now this is being exposed as we're using 17 digits. There is loss of precision, I believe due to double arithmetic, since we're passing in a very precise input (via the output file) here.

(cherry picked from commit 7ef960a39e6bbfd432dcb859d161ad2ed6086f97)