[ORCA] Support left/right outer join order hints
Commit ef7eec4 added a framework to support inner join order hints. Our next
step extends that framework to support left (LOJ) and right (ROJ) outer joins.

At a high level, join order hints is performed as a preprocessor step that
takes as input an nary join expression and a join order hint in order to
produce a new nary join expression composed of inner joins and LOJ/ROJs.

LOJ and ROJ are similar to inner joins, but have the following additional
restrictions:

Join predicates are non-transitive. For example:

    SELECT * FROM t1 LEFT JOIN t2 ON t1.a=42 LEFT JOIN t3 ON t1.a>4;

Predicates a=42 and a>4 cannot be pushed below their respective
join conditions. Otherwise NULL values may not be output correctly.

Join kind (e.g. left, right) is affected by the order. For example:

    SELECT * FROM t1 LEFT JOIN t2 ON t1.a=42;

    Leading((t1 t2)) requires a left join
    Leading((t2 t1)) requires a right join

In order to handle these additional constraints, the preprocessor step makes
decisions based on the LOJ specific attributes in CLogicalNAryJoin. In
particular, CScalarNAryJoinPredList.

Let's consider the following example:

    /*+
      Leading((t2 t1))
    */
    SELECT * FROM t t1 LEFT JOIN t t2 ON t1.a=t2.a LEFT JOIN t t3 ON t2.b=t3.b;

    Input:

    +--CLogicalNAryJoin [0, 1, 2]
       |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "b" (1), ...
       |--CLogicalGet "t2" ("t2"), Columns: ["a" (9), "b" (10), ...
       |--CLogicalGet "t3" ("t3"), Columns: ["a" (18), "b" (19), ...
       +--CScalarNAryJoinPredList
          |--CScalarConst (1)
          |--CScalarCmp (=)
          |  |--CScalarIdent "a" (0)
          |  +--CScalarIdent "a" (9)
          +--CScalarCmp (=)
             |--CScalarIdent "b" (10)
             +--CScalarIdent "b" (19)

The nary join's scalar child is either a scalar operator (if all the join
children are composed of inner joins) or a predicate list (if at least one join
child is a left outer join).

The nary join LOJ child predicate indexes [0, 1, 2] correspond to children t1,
t2, and t3. That means t1 left joins t2 using index 1 of LOJ child pred list
(i.e. "a"(0) = "a"(9) and that t2 left joins t3 using index 2 of LOJ child pred
list (i.e. "b"(10) = "b"(19).

In order to satisfy the hint, we need to convert a LOJ with t1 on the left side
to an ROJ with t1 on the right side. That requires maintaining a child "index"
so that we can compare the relative position of the child under the join. This
"index" is the value stored in the LOJ pred list index.

Relations [t1, t2] correspond to pred list indexes [0, 1]. The hint swaps the
order, so relations [t2, t1] correspond to pred list indexes [1, 0]. Note that
if the pred list values are ascending then it is a LOJ and if they are
descending then it is a ROJ. In this example, the hint produces a ROJ.

    Output:

    +--CLogicalNAryJoin [0, 1]
       |--CLogicalRightOuterJoin
       |  |--CLogicalGet "t2" ("t2"), Columns: ["a" (9), "b" (10), ...
       |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "b" (1), ...
       |  +--CScalarCmp (=)
       |     |--CScalarIdent "a" (0)
       |     +--CScalarIdent "a" (9)
       |--CLogicalGet "t3" ("t3"), Columns: ["a" (18), "b" (19), ...
       +--CScalarNAryJoinPredList
          |--CScalarConst (1)
          +--CScalarCmp (=)
             |--CScalarIdent "b" (10)
             +--CScalarIdent "b" (19)

After deciding to decompose a LOJ/ROJ, the pred list values need to be updated.
In the example, index 1 (i.e. "a"(0) = "a"(9)) is absorbed by the ROJ. That
changes the child pred list indexes from [0, 1, 2] to [0, 1]. In order to
account for this we need to track a set of used pred indexes. Here index 1 is
used so we need to subtract 1 from all indexes higher than 1.

    [0, 1, 2]  =>  [0, X, 2-1]  =>   [0, 1]
                       ^
                     deleted

(cherry picked from commit 568e0c340ff0916b10fc781fc7a08353ba889a1b)